requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aniso8601==8.0.0
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
mongoengine==0.20.0
multidict==4.7.6
pymongo==3.10.1
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
yarl==1.4.2
